Partial BorderLayer implementation
First visible drawing inside the control. BorderLayer is a child of
UILayer and it supports coloring, thickness and distance (how far from
the parent control's bounding box the border is drawn) properties.

Currently a border that's located at point 0,0 isn't drawn fully. It's
most likely a problem with some sort of internal padding, but requires
more investigation. Multiple borders can be stacked on top of each other
but the Z-index is not used to sort them yet.